🔀 :: (#464) 질문 삭제 API 구현

+++ src/main/java/kr/hs/entrydsm/rollsroyce/domain/qna/service/DeleteQnaService.java
@@ -1,4 +1,4 @@
-package kr.hs.entrydsm.rollsroyce.domain.admin.service;
+package kr.hs.entrydsm.rollsroyce.domain.qna.service;
 
 import lombok.RequiredArgsConstructor;
 
@@ -7,17 +7,26 @@ import org.springframework.transaction.annotation.Transactional;
 
 import kr.hs.entrydsm.rollsroyce.domain.qna.domain.Qna;
 import kr.hs.entrydsm.rollsroyce.domain.qna.domain.repository.QnaRepository;
+import kr.hs.entrydsm.rollsroyce.domain.qna.exception.QnaWriterMisMatchedException;
 import kr.hs.entrydsm.rollsroyce.domain.qna.facade.QnaFacade;
+import kr.hs.entrydsm.rollsroyce.domain.user.domain.User;
+import kr.hs.entrydsm.rollsroyce.domain.user.facade.UserFacade;
 
 @RequiredArgsConstructor
 @Service
 public class DeleteQnaService {
+    private final UserFacade userFacade;
     private final QnaFacade qnaFacade;
     private final QnaRepository qnaRepository;
 
     @Transactional
     public void execute(Long qnaId) {
+        User user = userFacade.getCurrentUser();
         Qna qna = qnaFacade.getQnaById(qnaId);
+
+        if (!user.getId().equals(qna.getUser().getId())) {
+            throw QnaWriterMisMatchedException.EXCEPTION;
+        }
 
         qnaRepository.delete(qna);
     }
